Add metadata support for collections
A better solution for metadata and collections in general would probably use
something like:

  `(into (quote ~empty-coll) ~cat)

where empty-coll is computed during macro expansion.  This would both preserve
metadata and be a gateway to less brittle support for all collection types.

But element order is reversed for lists and this complicates things beyond the
scope of this commit.